support generating files from Janet structs

+++ doers/src/file.rs
@@ -4,6 +4,7 @@ use blake3::Hash;
 use common::prelude::*;
 use regex::Regex;
 use serde::Deserialize;
+use serde_json::Value;
 use std::fmt::Debug;
 use std::fs;
 use std::io::Write;
@@ -28,6 +29,8 @@ pub struct DesiredFileState {
     pub content: Option<String>,
     pub ignore_pattern: Option<String>,
     pub from: Option<Utf8PathBuf>,
+    pub from_struct: Option<Value>,
+    pub to_format: Option<String>,
     pub backup_suffix: Option<String>,
 }
 
@@ -40,9 +43,16 @@ pub struct GurpFileRemove {
 }
 
 impl GurpFileEnsure {
-    fn content_xor_file(&self) -> bool {
-        (self.desired_state.content.is_some() && self.desired_state.from.is_none())
-            || (self.desired_state.content.is_none() && self.desired_state.from.is_some())
+    fn content_xor_file_xor_content_struct(&self) -> bool {
+        (self.desired_state.content.is_some()
+            && self.desired_state.from.is_none()
+            && self.desired_state.from_struct.is_none())
+            || (self.desired_state.content.is_none()
+                && self.desired_state.from.is_some()
+                && self.desired_state.from_struct.is_none())
+            || (self.desired_state.content.is_none()
+                && self.desired_state.from.is_none()
+                && self.desired_state.from_struct.is_some())
     }
 
     fn source_exists_if_needed(&self) -> bool {
@@ -60,6 +70,55 @@ impl GurpFileEnsure {
         }
     }
 
+    // Ini files can't nest structs. If we get anything we don't expect, error. This is very basic.
+    //
+    fn struct_to_ini(&self, value: &Value) -> anyhow::Result<String> {
+        let map = value
+            .as_object()
+            .ok_or_else(|| anyhow::anyhow!("Requested INI, but data is not a struct"))?;
+
+        let mut ret = String::new();
+
+        for (section_name, values) in map {
+            let section_map = values
+                .as_object()
+                .ok_or_else(|| anyhow::anyhow!("Section '{}' must be a struct", section_name))?;
+
+            if !ret.is_empty() {
+                ret.push('\n');
+            }
+
+            ret.push_str(&format!("[{section_name}]\n"));
+
+            for (k, v) in section_map {
+                let string_val = v.to_string().trim_matches('"').to_owned();
+                let value = if string_val.chars().all(|c| c.is_alphanumeric()) {
+                    string_val
+                } else {
+                    format!("\"{string_val}\"")
+                };
+
+                ret.push_str(&format!("{k} = {value}\n"));
+            }
+        }
+
+        Ok(ret)
+    }
+
+    fn struct_to_file(&self, value: &Value, format: Option<&str>) -> anyhow::Result<String> {
+        if let Some(format) = format {
+            match format.to_lowercase().as_str() {
+                "yaml" => Ok(serde_yaml_bw::to_string(&value)?),
+                "toml" => Ok(toml::to_string(&value)?),
+                "json" => Ok(serde_json::to_string_pretty(&value)?),
+                "ini" => Ok(self.struct_to_ini(value)?),
+                other => bail!("Unknown format: {}", other),
+            }
+        } else {
+            bail!("from_struct requires to_format")
+        }
+    }
+
     fn file_has_changed(&self) -> anyhow::Result<bool> {
         let (desired_hash, current_hash) = if let Some(pattern) = &self.desired_state.ignore_pattern
         {
@@ -68,6 +127,16 @@ impl GurpFileEnsure {
                     self.hash_of_filtered_file(from_file, pattern)?
                 } else if let Some(content) = &self.desired_state.content {
                     self.hash_of(&self.filter(content, pattern)?)
+                } else if let Some(from_struct) = &self.desired_state.from_struct {
+                    self.hash_of(
+                        &self.filter(
+                            &self.struct_to_file(
+                                from_struct,
+                                self.desired_state.to_format.as_deref(),
+                            )?,
+                            pattern,
+                        )?,
+                    )
                 } else {
                     bail!("No :content or :from")
                 },
@@ -79,6 +148,11 @@ impl GurpFileEnsure {
                     self.hash_of_file(from_file)?
                 } else if let Some(content) = &self.desired_state.content {
                     self.hash_of(content)
+                } else if let Some(from_struct) = &self.desired_state.from_struct {
+                    self.hash_of(
+                        &self
+                            .struct_to_file(from_struct, self.desired_state.to_format.as_deref())?,
+                    )
                 } else {
                     bail!("No :content or :from");
                 },
@@ -91,8 +165,8 @@ impl GurpFileEnsure {
 
     pub fn apply(&self, opts: &ApplyOpts) -> anyhow::Result<ApplySummary> {
         ensure!(
-            self.content_xor_file(),
-            "file '{}' must have exactly one of :content or :from",
+            self.content_xor_file_xor_content_struct(),
+            "file '{}' must have exactly one of :content, :from, or :from-struct",
             &self.path
         );
 
@@ -159,7 +233,11 @@ impl GurpFileEnsure {
 
         if let Some(from) = &self.desired_state.from {
             tracing::debug!("copying {} -> {}", from, self.path);
-            fs::copy(from, &self.path)?;
+
+            if !opts.noop {
+                fs::copy(from, &self.path)?;
+            }
+
             Ok(())
         } else if let Some(content) = &self.desired_state.content {
             tracing::debug!("Writing content to {}", self.path);
@@ -170,8 +248,20 @@ impl GurpFileEnsure {
             }
 
             Ok(())
+        } else if let Some(from_struct) = &self.desired_state.from_struct {
+            tracing::debug!("Writing file from struct to {}", self.path);
+
+            let content =
+                self.struct_to_file(from_struct, self.desired_state.to_format.as_deref())?;
+
+            if !opts.noop {
+                let mut fh = fs::File::create(&self.path)?;
+                fh.write_all(content.as_bytes())?;
+            }
+
+            Ok(())
         } else {
-            bail!("can write neither :from nor :content");
+            bail!("nothing to write. Require :from, :content, or :from-struct");
         }
     }
 
@@ -230,10 +320,11 @@ impl GurpFileRemove {
 #[cfg(test)]
 mod test {
     use super::*;
-    use assert_fs::prelude::*;
     use assert_fs::TempDir;
+    use assert_fs::prelude::*;
     use camino::Utf8PathBuf;
-    use indoc::formatdoc;
+    use indoc::{formatdoc, indoc};
+    use pretty_assertions::assert_eq;
     use std::fs;
     use std::os::unix::fs::PermissionsExt;
     use tester::{defopts, defopts_noop, fixture, janet2json, my_group, my_user};
@@ -243,13 +334,13 @@ mod test {
         let temp = TempDir::new().unwrap();
         let path = Utf8PathBuf::from_path_buf(temp.child("test-file").to_path_buf()).unwrap();
 
-        let json_def = janet2json(&formatdoc! {"
-            (file/ensure \"{}\"
-                :content \"some-junk\"
-                :mode \"0750\"
-                :owner \"{}\"
-                :group \"{}\")
-            ",
+        let json_def = janet2json(&formatdoc! {r#"
+            (file/ensure "{}"
+                :content "some-junk"
+                :mode "0750"
+                :owner "{}"
+                :group "{}")
+            "#,
             path,
             my_user(),
             my_group(),
@@ -268,13 +359,13 @@ mod test {
         let path = Utf8PathBuf::from_path_buf(file.to_path_buf()).unwrap();
         assert!(!path.exists());
 
-        let json_def = janet2json(&formatdoc! {"
-            (file/ensure \"{}\"
-                :content \"stuff\"
-                :mode \"0640\"
-                :owner \"{}\"
-                :group \"{}\")
-            ",
+        let json_def = janet2json(&formatdoc! {r#"
+            (file/ensure "{}"
+                :content "stuff"
+                :mode "0640"
+                :owner "{}"
+                :group "{}")
+            "#,
             path,
             my_user(),
             my_group(),
@@ -308,6 +399,8 @@ mod test {
                 ignore_pattern: None,
                 from: Some(fixture("file/binary-file")),
                 backup_suffix: None,
+                from_struct: None,
+                to_format: None,
             },
         };
 
@@ -333,6 +426,8 @@ mod test {
                 ignore_pattern: None,
                 from: Some(fixture("file/binary-file")),
                 backup_suffix: None,
+                from_struct: None,
+                to_format: None,
             },
         };
 
@@ -350,14 +445,14 @@ mod test {
         assert!(!path.exists());
 
         // escape { with another {
-        let json_def = janet2json(&formatdoc! {"
-            (file/ensure \"{}\"
-                :content (template-out \"{{{{ name }}}} is running a {{{{ thing }}}}\"
-                                        {{ :name \"gurp\" :thing \"test\" }})
-                :mode \"0600\"
-                :owner \"{}\"
-                :group \"{}\")
-            ",
+        let json_def = janet2json(&formatdoc! {r#"
+            (file/ensure "{}"
+                :content (template-out "{{{{ name }}}} is running a {{{{ thing }}}}"
+                                        {{ :name "gurp" :thing "test" }})
+                :mode "0600"
+                :owner "{}"
+                :group "{}")
+            "#,
             path,
             my_user(),
             my_group(),
@@ -372,6 +467,126 @@ mod test {
     }
 
     #[test]
+    fn test_file_create_json_from_struct() {
+        let temp = TempDir::new().unwrap();
+        let file = temp.join("test-file");
+        let path = Utf8PathBuf::from_path_buf(file.to_path_buf()).unwrap();
+        assert!(!path.exists());
+
+        let expected = indoc! { r#"
+                {
+                  "my-arr": [
+                    "abc",
+                    "def",
+                    "ghi"
+                  ],
+                  "my-str": "I am a String",
+                  "my-struct": {
+                    "key_1": "val 1",
+                    "key_2": 123,
+                    "key_3": [
+                      456,
+                      789
+                    ]
+                  }
+                }"#};
+
+        let sut: GurpFileEnsure = serde_json::from_str(&sample_struct(&path, "json")).unwrap();
+        assert_eq!(ONE_RESOURCE_ONE_CHANGE, sut.apply(&defopts()).unwrap());
+        assert!(path.exists());
+        let metadata = fs::metadata(&path).unwrap();
+        assert_eq!(metadata.permissions().mode() & 0o7777, 0o600);
+        assert_eq!(expected, fs::read_to_string(path).unwrap());
+    }
+
+    #[test]
+    fn test_file_create_yaml_from_struct() {
+        let temp = TempDir::new().unwrap();
+        let file = temp.join("test-file");
+        let path = Utf8PathBuf::from_path_buf(file.to_path_buf()).unwrap();
+        assert!(!path.exists());
+
+        let expected = indoc! { r#"
+            my-arr:
+            - abc
+            - def
+            - ghi
+            my-str: I am a String
+            my-struct:
+              key_1: val 1
+              key_2: 123
+              key_3:
+              - 456
+              - 789
+          "#};
+
+        let sut: GurpFileEnsure = serde_json::from_str(&sample_struct(&path, "yaml")).unwrap();
+        assert_eq!(ONE_RESOURCE_ONE_CHANGE, sut.apply(&defopts()).unwrap());
+        assert!(path.exists());
+        let metadata = fs::metadata(&path).unwrap();
+        assert_eq!(metadata.permissions().mode() & 0o7777, 0o600);
+        assert_eq!(expected, fs::read_to_string(path).unwrap());
+    }
+
+    #[test]
+    fn test_file_create_ini_from_struct() {
+        let temp = TempDir::new().unwrap();
+        let file = temp.join("test-file");
+        let path = Utf8PathBuf::from_path_buf(file.to_path_buf()).unwrap();
+        assert!(!path.exists());
+
+        let json_def = janet2json(&formatdoc! {r#"
+            (file/ensure "{path}"
+                :from-struct {{
+                    :section_1 {{
+                        :key_1 "A spacey string"
+                        :key_2 123
+                        :key_3 false
+                        :key_4 "word"
+                    }}
+                    :section_2 {{
+                        :key_1 "merp"
+                        :key_2 "gurp"
+                    }}
+                }}
+                :to-format "ini"
+                :mode "0600"
+                :owner "{user}"
+                :group "{group}")
+            "#,
+            path = path,
+            user = my_user(),
+            group = my_group(),
+        });
+
+        let expected = indoc! { r#"
+                [section_1]
+                key_1 = "A spacey string"
+                key_2 = 123
+                key_3 = false
+                key_4 = word
+
+                [section_2]
+                key_1 = merp
+                key_2 = gurp
+        "#};
+
+        let sut: GurpFileEnsure = serde_json::from_str(&json_def).unwrap();
+        assert_eq!(ONE_RESOURCE_ONE_CHANGE, sut.apply(&defopts()).unwrap());
+        assert!(path.exists());
+        let metadata = fs::metadata(&path).unwrap();
+        assert_eq!(metadata.permissions().mode() & 0o7777, 0o600);
+        assert_eq!(expected, fs::read_to_string(path).unwrap());
+    }
+
+    #[test]
+    fn test_file_create_ini_from_struct_errors() {
+        let sut: GurpFileEnsure =
+            serde_json::from_str(&sample_struct(&Utf8PathBuf::from("/tmp/file"), "ini")).unwrap();
+        assert!(sut.apply(&defopts()).is_err());
+    }
+
+    #[test]
     fn test_file_ensure_already_correct() {
         let temp = TempDir::new().unwrap();
         temp.child("test-file").write_str("stuff").unwrap();
@@ -381,13 +596,13 @@ mod test {
         fs::set_permissions(&path, fs::Permissions::from_mode(0o0750)).unwrap();
         assert!(path.exists());
 
-        let json_def = janet2json(&formatdoc! {"
-            (file/ensure \"{}\"
-                :content \"stuff\"
-                :mode \"0750\"
-                :owner \"{}\"
-                :group \"{}\")
-            ",
+        let json_def = janet2json(&formatdoc! {r#"
+            (file/ensure "{}"
+                :content "stuff"
+                :mode "0750"
+                :owner "{}"
+                :group "{}")
+            "#,
             path,
             my_user(),
             my_group(),
@@ -410,13 +625,13 @@ mod test {
             .join("test-file");
         assert!(path.exists());
 
-        let json_def = janet2json(&formatdoc! {"
-            (file/ensure \"{}\"
-                :content \"the-right-stuff\"
-                :mode \"0400\"
-                :owner \"{}\"
-                :group \"{}\")
-            ",
+        let json_def = janet2json(&formatdoc! {r#"
+            (file/ensure "{}"
+                :content "the-right-stuff"
+                :mode "0400"
+                :owner "{}"
+                :group "{}")
+            "#,
             path,
             my_user(),
             my_group(),
@@ -448,13 +663,13 @@ mod test {
 
         assert!(path.exists());
 
-        let json_def = janet2json(&formatdoc! {"
-            (file/ensure \"{}\"
-                :from \"{}\"
-                :mode \"0444\"
-                :owner \"{}\"
-                :group \"{}\")
-            ",
+        let json_def = janet2json(&formatdoc! {r#"
+            (file/ensure "{}"
+                :from "{}"
+                :mode "0444"
+                :owner "{}"
+                :group "{}")
+            "#,
             path,
             &fixture("file/copy-file"),
             my_user(),
@@ -518,14 +733,14 @@ mod test {
         fs::set_permissions(&path, fs::Permissions::from_mode(0o600)).unwrap();
         assert!(path.exists());
 
-        let json_def = janet2json(&formatdoc! {"
-            (file/ensure \"{}\"
-                :from \"{}\"
-                :mode \"0600\"
-                :ignore-pattern \"^today is\"
-                :owner \"{}\"
-                :group \"{}\")
-            ",
+        let json_def = janet2json(&formatdoc! {r#"
+            (file/ensure "{}"
+                :from "{}"
+                :mode "0600"
+                :ignore-pattern "^today is"
+                :owner "{}"
+                :group "{}")
+            "#,
             path,
             fixture("file/ignore-line-file"),
             my_user(),
@@ -585,5 +800,29 @@ mod test {
         let sut: GurpFileRemove = serde_json::from_str(&json_def).unwrap();
         assert_eq!(ONE_RESOURCE_NOOP, sut.apply(&defopts_noop()).unwrap());
         assert!(path.exists());
+    }
+
+    fn sample_struct(path: &Utf8PathBuf, format: &str) -> String {
+        janet2json(&formatdoc! {r#"
+            (file/ensure "{path}"
+                :from-struct {{
+                    :my-struct {{
+                        :key_1 "val 1"
+                        :key_2 123
+                        :key_3 [456 789]
+                    }}
+                    :my-arr ["abc" "def" "ghi"]
+                    :my-str "I am a String"
+                }}
+                :to-format "{format}"
+                :mode "0600"
+                :owner "{user}"
+                :group "{group}")
+            "#,
+            path = path,
+            format = format,
+            user = my_user(),
+            group = my_group(),
+        })
     }
 }
